mavlink::ardupilotmega

Struct EFI_STATUS_DATA

pub struct EFI_STATUS_DATA {

    pub ecu_index: f32,
    pub rpm: f32,
    pub fuel_consumed: f32,
    pub fuel_flow: f32,
    pub engine_load: f32,
    pub throttle_position: f32,
    pub spark_dwell_time: f32,
    pub barometric_pressure: f32,
    pub intake_manifold_pressure: f32,
    pub intake_manifold_temperature: f32,
    pub cylinder_head_temperature: f32,
    pub ignition_timing: f32,
    pub injection_time: f32,
    pub exhaust_gas_temperature: f32,
    pub throttle_out: f32,
    pub pt_compensation: f32,
    pub health: u8,
    pub ignition_voltage: f32,
    pub fuel_pressure: f32,

}

Available on crate feature ardupilotmega only.

EFI status output.

ID: 225

Fields
ecu_index: f32

ECU index

rpm: f32

RPM

fuel_consumed: f32

Fuel consumed

fuel_flow: f32

Fuel flow rate

engine_load: f32

Engine load

throttle_position: f32

Throttle position

spark_dwell_time: f32

Spark dwell time

barometric_pressure: f32

Barometric pressure

intake_manifold_pressure: f32

Intake manifold pressure(

intake_manifold_temperature: f32

Intake manifold temperature

cylinder_head_temperature: f32

Cylinder head temperature

ignition_timing: f32

Ignition timing (Crank angle degrees)

injection_time: f32

Injection time

exhaust_gas_temperature: f32

Exhaust gas temperature

throttle_out: f32

Output throttle

pt_compensation: f32

Pressure/temperature compensation

health: u8

EFI health status

ignition_voltage: f32

Supply voltage to EFI sparking system. Zero in this value means โunknownโ, so if the supply voltage really is zero volts use 0.0001 instead.

fuel_pressure: f32

Fuel pressure. Zero in this value means โunknownโ, so if the fuel pressure really is zero kPa use 0.0001 instead.

Implementations

impl EFI_STATUS_DATA

pub const ENCODED_LEN: usize = 73usize

pub const DEFAULT: Self