mavlink::ardupilotmega

Struct MISSION_CURRENT_DATA

pub struct MISSION_CURRENT_DATA {
    pub seq: u16,
    pub total: u16,
    pub mission_state: MissionState,
    pub mission_mode: u8,
    pub mission_id: u32,
    pub fence_id: u32,
    pub rally_points_id: u32,
}

Available on crate feature ardupilotmega only.

Message that announces the sequence number of the current target mission item (that the system will fly towards/execute when the mission is running). This message should be streamed all the time (nominally at 1Hz). This message should be emitted following a call to MAV_CMD_DO_SET_MISSION_CURRENT or MISSION_SET_CURRENT.

ID: 42

Fields
seq: u16

Sequence

total: u16

Total number of mission items on vehicle (on last item, sequence == total). If the autopilot stores its home location as part of the mission this will be excluded from the total. 0: Not supported, UINT16_MAX if no mission is present on the vehicle.

mission_state: MissionState

Mission state machine state. MISSION_STATE_UNKNOWN if state reporting not supported.

mission_mode: u8

Vehicle is in a mode that can execute mission items or suspended. 0: Unknown, 1: In mission mode, 2: Suspended (not in mission mode).

mission_id: u32

Id of current on-vehicle mission plan, or 0 if IDs are not supported or there is no mission loaded. GCS can use this to track changes to the mission plan type. The same value is returned on mission upload (in the MISSION_ACK).

fence_id: u32

Id of current on-vehicle fence plan, or 0 if IDs are not supported or there is no fence loaded. GCS can use this to track changes to the fence plan type. The same value is returned on fence upload (in the MISSION_ACK).

rally_points_id: u32

Id of current on-vehicle rally point plan, or 0 if IDs are not supported or there are no rally points loaded. GCS can use this to track changes to the rally point plan type. The same value is returned on rally point upload (in the MISSION_ACK).

Implementations

impl MISSION_CURRENT_DATA

pub const ENCODED_LEN: usize = 18usize

pub const DEFAULT: Self