mavlink::avssuas

Struct AUTOPILOT_VERSION_DATA

pub struct AUTOPILOT_VERSION_DATA {
    pub capabilities: MavProtocolCapability,
    pub uid: u64,
    pub flight_sw_version: u32,
    pub middleware_sw_version: u32,
    pub os_sw_version: u32,
    pub board_version: u32,
    pub vendor_id: u16,
    pub product_id: u16,
    pub flight_custom_version: [u8; 8],
    pub middleware_custom_version: [u8; 8],
    pub os_custom_version: [u8; 8],
    pub uid2: [u8; 18],
}

Available on crate feature avssuas only.

Version and capability of autopilot software. This should be emitted in response to a request with MAV_CMD_REQUEST_MESSAGE.

ID: 148

Fields
capabilities: MavProtocolCapability

Bitmap of capabilities

uid: u64

UID if provided by hardware (see uid2)

flight_sw_version: u32

Firmware version number. The field must be encoded as 4 bytes, where each byte (shown from MSB to LSB) is part of a semantic version: (major) (minor) (patch) (FIRMWARE_VERSION_TYPE).

middleware_sw_version: u32

Middleware version number

os_sw_version: u32

Operating system version number

board_version: u32

HW / board version (last 8 bits should be silicon ID, if any). The first 16 bits of this field specify https://github.com/PX4/PX4-Bootloader/blob/master/board_types.txt

vendor_id: u16

ID of the board vendor

product_id: u16

ID of the product

flight_custom_version: [u8; 8]

Custom version field, commonly the first 8 bytes of the git hash. This is not an unique identifier, but should allow to identify the commit using the main version number even for very large code bases.

middleware_custom_version: [u8; 8]

Custom version field, commonly the first 8 bytes of the git hash. This is not an unique identifier, but should allow to identify the commit using the main version number even for very large code bases.

os_custom_version: [u8; 8]

Custom version field, commonly the first 8 bytes of the git hash. This is not an unique identifier, but should allow to identify the commit using the main version number even for very large code bases.

uid2: [u8; 18]

UID if provided by hardware (supersedes the uid field. If this is non-zero, use this field, otherwise use uid)

Implementations

impl AUTOPILOT_VERSION_DATA

pub const ENCODED_LEN: usize = 78usize

pub const DEFAULT: Self