mavlink::avssuas

Struct GPS_STATUS_DATA

pub struct GPS_STATUS_DATA {
    pub satellites_visible: u8,
    pub satellite_prn: [u8; 20],
    pub satellite_used: [u8; 20],
    pub satellite_elevation: [u8; 20],
    pub satellite_azimuth: [u8; 20],
    pub satellite_snr: [u8; 20],
}

Available on crate feature avssuas only.

The positioning status, as reported by GPS. This message is intended to display status information about each satellite visible to the receiver. See message GLOBAL_POSITION_INT for the global position estimate. This message can contain information for up to 20 satellites.

ID: 25

Fields
satellites_visible: u8

Number of satellites visible

satellite_prn: [u8; 20]

Global satellite ID

satellite_used: [u8; 20]

0: Satellite not used, 1: used for localization

satellite_elevation: [u8; 20]

Elevation (0: right on top of receiver, 90: on the horizon) of satellite

satellite_azimuth: [u8; 20]

Direction of satellite, 0: 0 deg, 255: 360 deg.

satellite_snr: [u8; 20]

Signal to noise ratio of satellite

Implementations

impl GPS_STATUS_DATA

pub const ENCODED_LEN: usize = 101usize

pub const DEFAULT: Self