mavlink::avssuas

Struct SET_ATTITUDE_TARGET_DATA

pub struct SET_ATTITUDE_TARGET_DATA {
    pub time_boot_ms: u32,
    pub q: [f32; 4],
    pub body_roll_rate: f32,
    pub body_pitch_rate: f32,
    pub body_yaw_rate: f32,
    pub thrust: f32,
    pub target_system: u8,
    pub target_component: u8,
    pub type_mask: AttitudeTargetTypemask,
    pub thrust_body: [f32; 3],
}

Available on crate feature avssuas only.

Sets a desired vehicle attitude. Used by an external controller to command the vehicle (manual controller or other system).

ID: 82

Fields
time_boot_ms: u32

Timestamp (time since system boot).

q: [f32; 4]

Attitude quaternion (w, x, y, z order, zero-rotation is 1, 0, 0, 0) from MAV_FRAME_LOCAL_NED to MAV_FRAME_BODY_FRD

body_roll_rate: f32

Body roll rate

body_pitch_rate: f32

Body pitch rate

body_yaw_rate: f32

Body yaw rate

thrust: f32

Collective thrust, normalized to 0 .. 1 (-1 .. 1 for vehicles capable of reverse trust)

target_system: u8

System ID

target_component: u8

Component ID

type_mask: AttitudeTargetTypemask

Bitmap to indicate which dimensions should be ignored by the vehicle.

thrust_body: [f32; 3]

3D thrust setpoint in the body NED frame, normalized to -1 .. 1

Implementations

impl SET_ATTITUDE_TARGET_DATA

pub const ENCODED_LEN: usize = 51usize

pub const DEFAULT: Self