mavlink::cubepilot

Struct LINK_NODE_STATUS_DATA

pub struct LINK_NODE_STATUS_DATA {
    pub timestamp: u64,
    pub tx_rate: u32,
    pub rx_rate: u32,
    pub messages_sent: u32,
    pub messages_received: u32,
    pub messages_lost: u32,
    pub rx_parse_err: u16,
    pub tx_overflows: u16,
    pub rx_overflows: u16,
    pub tx_buf: u8,
    pub rx_buf: u8,
}

Available on crate feature cubepilot only.

Status generated in each node in the communication chain and injected into MAVLink stream.

ID: 8

Fields
timestamp: u64

Timestamp (time since system boot).

tx_rate: u32

Transmit rate

rx_rate: u32

Receive rate

messages_sent: u32

Messages sent

messages_received: u32

Messages received (estimated from counting seq)

messages_lost: u32

Messages lost (estimated from counting seq)

rx_parse_err: u16

Number of bytes that could not be parsed correctly.

tx_overflows: u16

Transmit buffer overflows. This number wraps around as it reaches UINT16_MAX

rx_overflows: u16

Receive buffer overflows. This number wraps around as it reaches UINT16_MAX

tx_buf: u8

Remaining free transmit buffer space

rx_buf: u8

Remaining free receive buffer space

Implementations

impl LINK_NODE_STATUS_DATA

pub const ENCODED_LEN: usize = 36usize

pub const DEFAULT: Self