mavlink::dialects::minimal

Enum MavMessage

#[repr(u32)]
pub enum MavMessage {
    HEARTBEAT(HEARTBEAT_DATA),
}

Available on crate feature dialect-minimal only.

Variants

HEARTBEAT(HEARTBEAT_DATA)

The heartbeat message shows that a system or component is present and responding. The type and autopilot fields (along with the message component id), allow the receiving system to treat further messages from this system appropriately (e.g. by laying out the user interface based on the autopilot). This microservice is documented at https://mavlink.io/en/services/heartbeat.html.

ID: 0

Implementations

impl MavMessage

pub const fn all_ids() -> &'static [u32]

pub const fn all_messages() -> &'static [(&'static str, u32)]