mavlink

Module minimal

Available on crate feature minimal only.

MAVLink minimal dialect.

This file was automatically generated, do not edit.

Structs

HEARTBEAT_DATA

The heartbeat message shows that a system or component is present and responding. The type and autopilot fields (along with the message component id), allow the receiving system to treat further messages from this system appropriately (e.g. by laying out the user interface based on the autopilot). This microservice is documented at https://mavlink.io/en/services/heartbeat.html.

MavModeFlag

These flags encode the MAV mode, see MAV_MODE enum for useful combinations.

Enums

MavAutopilot

Micro air vehicle / autopilot classes. This identifies the individual model.

MavComponent

Legacy component ID values for particular types of hardware/software that might make up a MAVLink system (autopilot, cameras, servos, avoidance systems etc.). Components are not required or expected to use IDs with names that correspond to their type or function, but may choose to do so. Using an ID that matches the type may slightly reduce the chances of component id clashes, as, for historical reasons, it is less likely to be used by some other type of component. System integration will still need to ensure that all components have unique IDs. Component IDs are used for addressing messages to a particular component within a system. A component can use any unique ID between 1 and 255 (MAV_COMP_ID_ALL value is the broadcast address, used to send to all components). Historically component ID were also used for identifying the type of component. New code must not use component IDs to infer the component type, but instead check the MAV_TYPE in the HEARTBEAT message!

MavMessage

MavModeFlagDecodePosition

These values encode the bit positions of the decode position. These values can be used to read the value of a flag bit by combining the base_mode variable with AND with the flag position value. The result will be either 0 or 1, depending on if the flag is set or not.

MavState

MavType

MAVLINK component type reported in HEARTBEAT message. Flight controllers must report the type of the vehicle on which they are mounted (e.g. MAV_TYPE_OCTOROTOR). All other components must report a value appropriate for their type (e.g. a camera must use MAV_TYPE_CAMERA).

Constants

MINOR_MAVLINK_VERSION