mavlink::minimal

Enum MavType

#[repr(u32)]
pub enum MavType {

    MAV_TYPE_GENERIC = 0,
    MAV_TYPE_FIXED_WING = 1,
    MAV_TYPE_QUADROTOR = 2,
    MAV_TYPE_COAXIAL = 3,
    MAV_TYPE_HELICOPTER = 4,
    MAV_TYPE_ANTENNA_TRACKER = 5,
    MAV_TYPE_GCS = 6,
    MAV_TYPE_AIRSHIP = 7,
    MAV_TYPE_FREE_BALLOON = 8,
    MAV_TYPE_ROCKET = 9,
    MAV_TYPE_GROUND_ROVER = 10,
    MAV_TYPE_SURFACE_BOAT = 11,
    MAV_TYPE_SUBMARINE = 12,
    MAV_TYPE_HEXAROTOR = 13,
    MAV_TYPE_OCTOROTOR = 14,
    MAV_TYPE_TRICOPTER = 15,
    MAV_TYPE_FLAPPING_WING = 16,
    MAV_TYPE_KITE = 17,
    MAV_TYPE_ONBOARD_CONTROLLER = 18,
    MAV_TYPE_VTOL_TAILSITTER_DUOROTOR = 19,
    MAV_TYPE_VTOL_TAILSITTER_QUADROTOR = 20,
    MAV_TYPE_VTOL_TILTROTOR = 21,
    MAV_TYPE_VTOL_FIXEDROTOR = 22,
    MAV_TYPE_VTOL_TAILSITTER = 23,
    MAV_TYPE_VTOL_TILTWING = 24,
    MAV_TYPE_VTOL_RESERVED5 = 25,
    MAV_TYPE_GIMBAL = 26,
    MAV_TYPE_ADSB = 27,
    MAV_TYPE_PARAFOIL = 28,
    MAV_TYPE_DODECAROTOR = 29,
    MAV_TYPE_CAMERA = 30,
    MAV_TYPE_CHARGING_STATION = 31,
    MAV_TYPE_FLARM = 32,
    MAV_TYPE_SERVO = 33,
    MAV_TYPE_ODID = 34,
    MAV_TYPE_DECAROTOR = 35,
    MAV_TYPE_BATTERY = 36,
    MAV_TYPE_PARACHUTE = 37,
    MAV_TYPE_LOG = 38,
    MAV_TYPE_OSD = 39,
    MAV_TYPE_IMU = 40,
    MAV_TYPE_GPS = 41,
    MAV_TYPE_WINCH = 42,
    MAV_TYPE_GENERIC_MULTIROTOR = 43,
    MAV_TYPE_ILLUMINATOR = 44,
    MAV_TYPE_SPACECRAFT_ORBITER = 45,

}

Available on crate feature minimal only.

MAVLINK component type reported in HEARTBEAT message. Flight controllers must report the type of the vehicle on which they are mounted (e.g. MAV_TYPE_OCTOROTOR). All other components must report a value appropriate for their type (e.g. a camera must use MAV_TYPE_CAMERA).

Variants

MAV_TYPE_GENERIC = 0

Generic micro air vehicle

MAV_TYPE_FIXED_WING = 1

Fixed wing aircraft.

MAV_TYPE_QUADROTOR = 2

Quadrotor

MAV_TYPE_COAXIAL = 3

Coaxial helicopter

MAV_TYPE_HELICOPTER = 4

Normal helicopter with tail rotor.

MAV_TYPE_ANTENNA_TRACKER = 5

Ground installation

MAV_TYPE_GCS = 6

Operator control unit / ground control station

MAV_TYPE_AIRSHIP = 7

Airship, controlled

MAV_TYPE_FREE_BALLOON = 8

Free balloon, uncontrolled

MAV_TYPE_ROCKET = 9

Rocket

MAV_TYPE_GROUND_ROVER = 10

Ground rover

MAV_TYPE_SURFACE_BOAT = 11

Surface vessel, boat, ship

MAV_TYPE_SUBMARINE = 12

Submarine

MAV_TYPE_HEXAROTOR = 13

Hexarotor

MAV_TYPE_OCTOROTOR = 14

Octorotor

MAV_TYPE_TRICOPTER = 15

Tricopter

MAV_TYPE_FLAPPING_WING = 16

Flapping wing

MAV_TYPE_KITE = 17

Kite

MAV_TYPE_ONBOARD_CONTROLLER = 18

Onboard companion controller

MAV_TYPE_VTOL_TAILSITTER_DUOROTOR = 19

Two-rotor Tailsitter VTOL that additionally uses control surfaces in vertical operation. Note, value previously named MAV_TYPE_VTOL_DUOROTOR.

MAV_TYPE_VTOL_TAILSITTER_QUADROTOR = 20

Quad-rotor Tailsitter VTOL using a V-shaped quad config in vertical operation. Note: value previously named MAV_TYPE_VTOL_QUADROTOR.

MAV_TYPE_VTOL_TILTROTOR = 21

Tiltrotor VTOL. Fuselage and wings stay (nominally) horizontal in all flight phases. It able to tilt (some) rotors to provide thrust in cruise flight.

MAV_TYPE_VTOL_FIXEDROTOR = 22

VTOL with separate fixed rotors for hover and cruise flight. Fuselage and wings stay (nominally) horizontal in all flight phases.

MAV_TYPE_VTOL_TAILSITTER = 23

Tailsitter VTOL. Fuselage and wings orientation changes depending on flight phase: vertical for hover, horizontal for cruise. Use more specific VTOL MAV_TYPE_VTOL_TAILSITTER_DUOROTOR or MAV_TYPE_VTOL_TAILSITTER_QUADROTOR if appropriate.

MAV_TYPE_VTOL_TILTWING = 24

Tiltwing VTOL. Fuselage stays horizontal in all flight phases. The whole wing, along with any attached engine, can tilt between vertical and horizontal mode.

MAV_TYPE_VTOL_RESERVED5 = 25

VTOL reserved 5

MAV_TYPE_GIMBAL = 26

Gimbal

MAV_TYPE_ADSB = 27

ADSB system

MAV_TYPE_PARAFOIL = 28

Steerable, nonrigid airfoil

MAV_TYPE_DODECAROTOR = 29

Dodecarotor

MAV_TYPE_CAMERA = 30

Camera

MAV_TYPE_CHARGING_STATION = 31

Charging station

MAV_TYPE_FLARM = 32

FLARM collision avoidance system

MAV_TYPE_SERVO = 33

Servo

MAV_TYPE_ODID = 34

Open Drone ID. See https://mavlink.io/en/services/opendroneid.html.

MAV_TYPE_DECAROTOR = 35

Decarotor

MAV_TYPE_BATTERY = 36

Battery

MAV_TYPE_PARACHUTE = 37

Parachute

MAV_TYPE_LOG = 38

Log

MAV_TYPE_OSD = 39

OSD

MAV_TYPE_IMU = 40

IMU

MAV_TYPE_GPS = 41

GPS

MAV_TYPE_WINCH = 42

Winch

MAV_TYPE_GENERIC_MULTIROTOR = 43

Generic multirotor that does not fit into a specific type or whose type is unknown

MAV_TYPE_ILLUMINATOR = 44

Illuminator. An illuminator is a light source that is used for lighting up dark areas external to the sytstem: e.g. a torch or searchlight (as opposed to a light source for illuminating the system itself, e.g. an indicator light).

MAV_TYPE_SPACECRAFT_ORBITER = 45

Orbiter spacecraft. Includes satellites orbiting terrestrial and extra-terrestrial bodies. Follows NASA Spacecraft Classification.

Implementations

impl MavType

pub const DEFAULT: Self = Self::MAV_TYPE_GENERIC