mavlink::python_array_test

Struct POSITION_TARGET_GLOBAL_INT_DATA

pub struct POSITION_TARGET_GLOBAL_INT_DATA {

    pub time_boot_ms: u32,
    pub lat_int: i32,
    pub lon_int: i32,
    pub alt: f32,
    pub vx: f32,
    pub vy: f32,
    pub vz: f32,
    pub afx: f32,
    pub afy: f32,
    pub afz: f32,
    pub yaw: f32,
    pub yaw_rate: f32,
    pub type_mask: PositionTargetTypemask,
    pub coordinate_frame: MavFrame,

}

Available on crate feature python_array_test only.

id: 87 Reports the current commanded vehicle position, velocity, and acceleration as specified by the autopilot. This should match the commands sent in SET_POSITION_TARGET_GLOBAL_INT if the vehicle is being controlled this way.

Fields
time_boot_ms: u32

Timestamp (time since system boot). The rationale for the timestamp in the setpoint is to allow the system to compensate for the transport delay of the setpoint. This allows the system to compensate processing latency.

lat_int: i32

Latitude in WGS84 frame

lon_int: i32

Longitude in WGS84 frame

alt: f32

Altitude (MSL, AGL or relative to home altitude, depending on frame)

vx: f32

X velocity in NED frame

vy: f32

Y velocity in NED frame

vz: f32

Z velocity in NED frame

afx: f32

X acceleration or force (if bit 10 of type_mask is set) in NED frame in meter / s^2 or N

afy: f32

Y acceleration or force (if bit 10 of type_mask is set) in NED frame in meter / s^2 or N

afz: f32

Z acceleration or force (if bit 10 of type_mask is set) in NED frame in meter / s^2 or N

yaw: f32

yaw setpoint

yaw_rate: f32

yaw rate setpoint

type_mask: PositionTargetTypemask

Bitmap to indicate which dimensions should be ignored by the vehicle.

coordinate_frame: MavFrame

Valid options are: MAV_FRAME_GLOBAL = 0, MAV_FRAME_GLOBAL_RELATIVE_ALT = 3, MAV_FRAME_GLOBAL_TERRAIN_ALT = 10 (MAV_FRAME_GLOBAL_INT, MAV_FRAME_GLOBAL_RELATIVE_ALT_INT, MAV_FRAME_GLOBAL_TERRAIN_ALT_INT are allowed synonyms, but have been deprecated)

Implementations

impl POSITION_TARGET_GLOBAL_INT_DATA

pub const ENCODED_LEN: usize = 51usize

pub const DEFAULT: Self