mavlink

Struct MAVLinkV1MessageRaw

pub struct MAVLinkV1MessageRaw(/* private fields */);

Byte buffer containing the raw representation of a MAVLink 1 message beginning with the STX marker.

Follow protocol definition: https://mavlink.io/en/guide/serialization.html#v1_packet_format. Maximum size is 263 bytes.

Implementations

impl MAVLinkV1MessageRaw

pub const fn new() -> MAVLinkV1MessageRaw

Create a new raw MAVLink 1 message filled with zeros.

pub const fn from_bytes_unparsed(bytes: [u8; 263]) -> MAVLinkV1MessageRaw

Create a new raw MAVLink 1 message from a given buffer.

Note: This method does not guarantee that the constructed MAVLink message is valid.

pub fn as_slice(&self) -> &[u8]

Read access to its internal buffer.

pub fn as_mut_slice(&mut self) -> &mut [u8]

Mutable reference to its internal buffer.

pub fn into_inner(self) -> [u8; 263]

Deconstruct the MAVLink message into its owned internal buffer.

pub fn header(&self) -> &[u8]

Reference to the 5 byte header slice of the message

pub fn payload_length(&self) -> u8

Size of the payload of the message

pub fn sequence(&self) -> u8

Packet sequence number

pub fn system_id(&self) -> u8

Message sender System ID

pub fn component_id(&self) -> u8

Message sender Component ID

pub fn message_id(&self) -> u8

Message ID

pub fn payload(&self) -> &[u8]

Reference to the payload byte slice of the message

pub fn checksum(&self) -> u16

CRC-16 checksum field of the message

pub fn has_valid_crc<M>(&self) -> bool
where M: Message,

Checks wether the messageโs CRC-16 checksum calculation matches its checksum field.

pub fn raw_bytes(&self) -> &[u8]

Raw byte slice of the message

pub fn serialize_message<M>(&mut self, header: MavHeader, message: &M)
where M: Message,

Serialize a Message with a given header into this raw message buffer.

Panics

If the messageโs id exceeds 255 and is therefore not supported for MAVLink 1

pub fn serialize_message_data<D>(&mut self, header: MavHeader, message_data: &D)
where D: MessageData,

Panics

If the MessageDataโs ID exceeds 255 and is therefore not supported for MAVLink 1