mavlink

Trait Connectable

pub trait Connectable: Display {
    // Required method
    fn connect<M>(
        &self,
    ) -> Result<Box<dyn MavConnection<M> + Send + Sync>, Error>
       where M: Message;
}

A MAVLink connection address that can be connected to, establishing a MavConnection

Required Methods

fn connect<M>(&self) -> Result<Box<dyn MavConnection<M> + Send + Sync>, Error>
where M: Message,

Attempt to establish a blocking MAVLink connection

Errors

When the connection could not be established a corresponding io::Error is returned

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.