mavlink::ualberta

Struct SIM_STATE_DATA

pub struct SIM_STATE_DATA {

    pub q1: f32,
    pub q2: f32,
    pub q3: f32,
    pub q4: f32,
    pub roll: f32,
    pub pitch: f32,
    pub yaw: f32,
    pub xacc: f32,
    pub yacc: f32,
    pub zacc: f32,
    pub xgyro: f32,
    pub ygyro: f32,
    pub zgyro: f32,
    pub lat: f32,
    pub lon: f32,
    pub alt: f32,
    pub std_dev_horz: f32,
    pub std_dev_vert: f32,
    pub vn: f32,
    pub ve: f32,
    pub vd: f32,
    pub lat_int: i32,
    pub lon_int: i32,

}

Available on crate feature ualberta only.

Status of simulation environment, if used.

ID: 108

Fields
q1: f32

True attitude quaternion component 1, w (1 in null-rotation)

q2: f32

True attitude quaternion component 2, x (0 in null-rotation)

q3: f32

True attitude quaternion component 3, y (0 in null-rotation)

q4: f32

True attitude quaternion component 4, z (0 in null-rotation)

roll: f32

Attitude roll expressed as Euler angles, not recommended except for human-readable outputs

pitch: f32

Attitude pitch expressed as Euler angles, not recommended except for human-readable outputs

yaw: f32

Attitude yaw expressed as Euler angles, not recommended except for human-readable outputs

xacc: f32

X acceleration

yacc: f32

Y acceleration

zacc: f32

Z acceleration

xgyro: f32

Angular speed around X axis

ygyro: f32

Angular speed around Y axis

zgyro: f32

Angular speed around Z axis

lat: f32

Latitude (lower precision). Both this and the lat_int field should be set.

lon: f32

Longitude (lower precision). Both this and the lon_int field should be set.

alt: f32

Altitude

std_dev_horz: f32

Horizontal position standard deviation

std_dev_vert: f32

Vertical position standard deviation

vn: f32

True velocity in north direction in earth-fixed NED frame

ve: f32

True velocity in east direction in earth-fixed NED frame

vd: f32

True velocity in down direction in earth-fixed NED frame

lat_int: i32

Latitude (higher precision). If 0, recipients should use the lat field value (otherwise this field is preferred).

lon_int: i32

Longitude (higher precision). If 0, recipients should use the lon field value (otherwise this field is preferred).

Implementations

impl SIM_STATE_DATA

pub const ENCODED_LEN: usize = 92usize

pub const DEFAULT: Self