mavlink::uavionix

Struct COMMAND_ACK_DATA

pub struct COMMAND_ACK_DATA {
    pub command: MavCmd,
    pub result: MavResult,
    pub progress: u8,
    pub result_param2: i32,
    pub target_system: u8,
    pub target_component: u8,
}

Available on crate feature uavionix only.

Report status of a command. Includes feedback whether the command was executed. The command microservice is documented at https://mavlink.io/en/services/command.html.

ID: 77

Fields
command: MavCmd

Command ID (of acknowledged command).

result: MavResult

Result of command.

progress: u8

The progress percentage when result is MAV_RESULT_IN_PROGRESS. Values: [0-100], or UINT8_MAX if the progress is unknown.

result_param2: i32

Additional result information. Can be set with a command-specific enum containing command-specific error reasons for why the command might be denied. If used, the associated enum must be documented in the corresponding MAV_CMD (this enum should have a 0 value to indicate โunusedโ or โunknownโ).

target_system: u8

System ID of the target recipient. This is the ID of the system that sent the command for which this COMMAND_ACK is an acknowledgement.

target_component: u8

Component ID of the target recipient. This is the ID of the system that sent the command for which this COMMAND_ACK is an acknowledgement.

Implementations

impl COMMAND_ACK_DATA

pub const ENCODED_LEN: usize = 10usize

pub const DEFAULT: Self