Crate mavlink

Modules

ardupilotmega

This file was automatically generated, do not edit

bytes

bytes_mut

common

This file was automatically generated, do not edit

error

icarous

This file was automatically generated, do not edit

peek_reader

This module implements a buffered/peekable reader.

uavionix

This file was automatically generated, do not edit

utils

Structs

MAVLinkV1MessageRaw

MAVLinkV2MessageRaw

MavFrame

Encapsulation of the Mavlink message and the header, important to preserve information about the sender system and component id

MavHeader

Metadata from a MAVLink packet header

Enums

MavlinkVersion

Versions of the Mavlink protocol that we support

Constants

MAV_STX

Message framing marker for mavlink v1

MAV_STX_V2

Message framing marker for mavlink v2

MAX_FRAME_SIZE

Traits

MavConnection

A MAVLink connection

Message

MessageData

Functions

calculate_crc

connect

Connect to a MAVLink node by address string.

read_v1_msg

Read a MAVLink v1 message from a Read stream.

read_v1_raw_message

Return a raw buffer with the mavlink message V1 maximum size is 263 bytes: <https://mavlink.io/en/guide/serialization.html>

read_v2_msg

Read a MAVLink v2 message from a Read stream.

read_v2_raw_message

Return a raw buffer with the mavlink message V2 maximum size is 280 bytes: <https://mavlink.io/en/guide/serialization.html>

read_versioned_msg

write_v1_msg

Write a MAVLink v1 message to a Write stream.

write_v2_msg

Write a MAVLink v2 message to a Write stream.

write_versioned_msg

Write a message using the given mavlink version